refactor: unified cube datatset under the SyntheticScene class. Removed useless and redondant test
